pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-explained-with-code"
version = "0.1.0"
description = "A Retrieval-Augmented Generation (RAG) product recommendation assistant"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Praveen Kannan", email = "praveen.kannan@example.com" }
]
dependencies = [
    "openai>=1.0.0",
    "numpy>=1.22.0",
    "python-dotenv>=0.21.0",
    "faiss-cpu>=1.7.4",
    "tiktoken>=0.5.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "mypy>=1.3.0",
    "ruff>=0.0.272"
]

[project.scripts]
product-assistant = "product_assistant:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "-v"
]
